Refactorin folders and add types Request of express
