+++ src/@types/express/index.d.ts
@@ -1,0 +1,8 @@
+declare namespace Express {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    export interface Request {
+        user: {
+            id: string;
+        };
+    }
+}
