+++ tsconfig.json
@@ -7,7 +7,7 @@
     "outDir": "./dist",
     "paths": {
       "@modules/*": ["modules/*"],
-      "@config/*": ["config/*"],
+      "@config/*": ["src/config/*"],
       "@shared/*": ["shared/*"],
       "@errors/*": ["errors/*"],
       "@utils/*": ["utils/*"]
